Move docgen into krpctools
