--- tools/docgen/README.txt
@@ -1,1 +1,0 @@
-Tool for generating documentation for kRPC
