+++ tools/krpctools/README.txt
@@ -1,1 +1,5 @@
-Tool to generate client stub code for kRPC services
+Tools for kRPC:
+
+ - krpc-clientgen : generate client stub code for kRPC services
+ - krpc-docgen : generate documentation for kRPC services
+ - krpc-servicedefs : generate service definition JSON files for kRPC services
