+++ tools/krpctools/docgen.bzl
@@ -29,7 +29,7 @@ docgen = rule(
         'language': attr.string(mandatory=True),
         'src': attr.label(allow_files=True, single_file=True),
         'defs': attr.label_list(allow_files=True, mandatory=True, non_empty=True),
-        '_docgen': attr.label(default=Label('//tools/docgen:script'), executable=True, allow_files=True, single_file=True),
+        '_docgen': attr.label(default=Label('//tools/krpctools:docgen'), executable=True, allow_files=True, single_file=True),
         '_order': attr.label(default=Label('//doc:order.txt'), allow_files=True, single_file=True)
     },
     outputs = _outputs
