+++ tools/krpctools/krpctools/docgen/lua.py
@@ -1,60 +1,73 @@
-from krpc.docgen.domain import Domain
-from krpc.docgen.nodes import *
-from krpc.docgen.utils import snakecase
+from .domain import Domain
+from .nodes import *
+from .utils import snakecase
 from krpc.types import ValueType, ClassType, EnumType, ListType, DictionaryType, SetType, TupleType
 
-class PythonDomain(Domain):
-    name = 'python'
-    prettyname = 'Python'
-    sphinxname = 'py'
-    codeext = 'py'
+class LuaDomain(Domain):
+
+    name = 'lua'
+    prettyname = 'Lua'
+    sphinxname = 'lua'
+    codeext = 'lua'
+
+    type_map = {
+        'double': 'number',
+        'float': 'number',
+        'int32': 'number',
+        'int64': 'number',
+        'uint32': 'number',
+        'uint64': 'number',
+        'bool': 'boolean',
+        'string': 'string',
+        'bytes': 'string'
+    }
 
     value_map = {
-        'null': 'None',
+        'null': 'nil',
         'true': 'True',
         'false': 'False'
     }
 
     def __init__(self, macros):
-        super(PythonDomain, self).__init__(macros)
+        super(LuaDomain, self).__init__(macros)
 
     def currentmodule(self, name):
-        super(PythonDomain, self).currentmodule(name)
+        super(LuaDomain, self).currentmodule(name)
         return '.. currentmodule:: %s' % name
 
     def type(self, typ):
         if isinstance(typ, ValueType):
-            return typ.python_type.__name__
+            return self.type_map[typ.protobuf_type]
         elif isinstance(typ, ClassType):
             return self.shorten_ref(typ.protobuf_type[6:-1])
         elif isinstance(typ, EnumType):
             return self.shorten_ref(typ.protobuf_type[5:-1])
         elif isinstance(typ, ListType):
-            return 'list'
+            return 'List'
         elif isinstance(typ, DictionaryType):
-            return 'dict'
+            return 'Map'
         elif isinstance(typ, SetType):
-            return 'set'
+            return 'Set'
         elif isinstance(typ, TupleType):
-            return 'tuple'
+            return 'Tuple'
         else:
             raise RuntimeError('Unknown type \'%s\'' % str(typ))
 
     def type_description(self, typ):
         if isinstance(typ, ValueType):
-            return typ.python_type.__name__
+            return self.type_map[typ.protobuf_type]
         elif isinstance(typ, ClassType):
             return ':class:`%s`' % self.type(typ)
         elif isinstance(typ, EnumType):
             return ':class:`%s`' % self.type(typ)
         elif isinstance(typ, ListType):
-            return 'list of %s' % self.type_description(typ.value_type)
+            return 'List of %s' % self.type_description(typ.value_type)
         elif isinstance(typ, DictionaryType):
-            return 'dict from %s to %s' % (self.type_description(typ.key_type), self.type_description(typ.value_type))
+            return 'Map from %s to %s' % (self.type_description(typ.key_type), self.type_description(typ.value_type))
         elif isinstance(typ, SetType):
-            return 'set of %s' % self.type_description(typ.value_type)
+            return 'Set of %s' % self.type_description(typ.value_type)
         elif isinstance(typ, TupleType):
-            return 'tuple of (%s)' % ', '.join(self.type_description(typ) for typ in typ.value_types)
+            return 'Tuple of (%s)' % ', '.join(self.type_description(typ) for typ in typ.value_types)
         else:
             raise RuntimeError('Unknown type \'%s\'' % str(typ))
 
@@ -80,4 +93,4 @@ class PythonDomain(Domain):
         return ':%s:`%s`' % (prefix, self.ref(obj))
 
     def paramref(self, name):
-        return super(PythonDomain, self).paramref(snakecase(name))
+        return super(LuaDomain, self).paramref(snakecase(name))
