+++ tools/krpctools/krpctools/docgen/utils.py
